Move GetFeeds to service layer
